Create LitTrait and remove repetitive code
